reusable MdxEditor component
